change read angle from calib_servo.noexiit to fl

+++ software/pt_to_pt_stream_expt.py
@@ -118,7 +118,7 @@ def main():
     if ext_angle is None:
         with open ("calib_servo.noexiit", "r") as f:
             max_ext = f.read().rstrip('\n')
-        ext_angle = int(max_ext)
+        ext_angle = float(max_ext)
     
     # Set up filename to save:
     t_script_start = datetime.datetime.now()
